Added new user and delete user
